added incoherent prng for terrain gen (no more GD.Randf()) and increased the wackitude of terrain generation

+++ src/util/LayeredNoise.cs
@@ -16,9 +16,9 @@ public class LayeredNoise
     public int Seed { get; }
     private float variance;
 
-    public LayeredNoise(int seed = 1337)
+    public LayeredNoise(uint seed = 1337)
     {
-        Seed = seed;
+        Seed = (int)seed;
     }
 
     public float Scale => sumNoiseMagnitude;
